move enum to enum utils
