+++ deepfence_frontend/apps/dashboard/src/components/SeverityBadge.tsx
@@ -15,7 +15,7 @@ import {
 } from '@/constants/charts';
 import { Mode, useTheme } from '@/theme/ThemeContext';
 import { PostureSeverityType, VulnerabilitySeverityType } from '@/types/common';
-import { SeverityEnum } from '@/utils/scan';
+import { SeverityEnum } from '@/utils/enum';
 
 export const SeverityBadge = ({
   severity,
