+++ deepfence_frontend/apps/dashboard/src/features/malwares/components/landing/TopNMalwareCard.tsx
@@ -13,8 +13,8 @@ import { getSeverityColorMap } from '@/constants/charts';
 import { CardHeader } from '@/features/malwares/pages/landing/CardHeader';
 import { queries } from '@/queries';
 import { Mode, useTheme } from '@/theme/ThemeContext';
+import { SeverityEnum } from '@/utils/enum';
 import { abbreviateNumber } from '@/utils/number';
-import { SeverityEnum } from '@/utils/scan';
 import { usePageNavigation } from '@/utils/usePageNavigation';
 
 export interface TopNMalwareChartData {
