+++ deepfence_frontend/apps/dashboard/src/features/malwares/pages/MostExploitableMalwares.tsx
@@ -34,7 +34,7 @@ import { FilterWrapper } from '@/features/common/FilterWrapper';
 import { queries } from '@/queries';
 import { useTheme } from '@/theme/ThemeContext';
 import { MalwareSeverityType, ScanTypeEnum } from '@/types/common';
-import { SeverityEnumList } from '@/utils/scan';
+import { SeverityEnumList } from '@/utils/enum';
 
 const DEFAULT_PAGE_SIZE = 10;
