+++ deepfence_frontend/apps/dashboard/src/features/malwares/pages/UniqueMalwares.tsx
@@ -34,7 +34,7 @@ import { FilterWrapper } from '@/features/common/FilterWrapper';
 import { queries } from '@/queries';
 import { useTheme } from '@/theme/ThemeContext';
 import { MalwareSeverityType, ScanTypeEnum } from '@/types/common';
-import { SeverityEnumList } from '@/utils/scan';
+import { SeverityEnumList } from '@/utils/enum';
 import { getOrderFromSearchParams, useSortingState } from '@/utils/table';
 
 const DEFAULT_PAGE_SIZE = 10;
