+++ deepfence_frontend/apps/dashboard/src/features/onboard/pages/VulnerabilityScanSummary.tsx
@@ -10,7 +10,7 @@ import { ConnectorHeader } from '@/features/onboard/components/ConnectorHeader';
 import { queries } from '@/queries';
 import { useTheme } from '@/theme/ThemeContext';
 import { ScanTypeEnum } from '@/types/common';
-import { SeverityEnum } from '@/utils/scan';
+import { SeverityEnum } from '@/utils/enum';
 
 const DEFAULT_PAGE_SIZE = 10;
