+++ deepfence_frontend/apps/dashboard/src/features/secrets/pages/MostExploitableSecrets.tsx
@@ -33,7 +33,7 @@ import { FilterWrapper } from '@/features/common/FilterWrapper';
 import { queries } from '@/queries';
 import { useTheme } from '@/theme/ThemeContext';
 import { ScanTypeEnum, SecretSeverityType } from '@/types/common';
-import { SeverityEnumList } from '@/utils/scan';
+import { SeverityEnumList } from '@/utils/enum';
 
 const DEFAULT_PAGE_SIZE = 10;
