+++ deepfence_frontend/apps/dashboard/src/features/vulnerabilities/pages/MostExploitableVulnerabilities.tsx
@@ -39,11 +39,7 @@ import { DownloadLineIcon } from '@/components/icons/common/DownloadLine';
 import { FilterIcon } from '@/components/icons/common/Filter';
 import { PopOutIcon } from '@/components/icons/common/PopOut';
 import { TimesIcon } from '@/components/icons/common/Times';
-import {
-  CveCVSSScore,
-  SeverityBadge,
-  SeverityBadgeIcon,
-} from '@/components/SeverityBadge';
+import { SeverityBadgeIcon } from '@/components/SeverityBadge';
 import { VulnerabilityIcon } from '@/components/sideNavigation/icons/Vulnerability';
 import { TruncatedText } from '@/components/TruncatedText';
 import { BreadcrumbWrapper } from '@/features/common/BreadcrumbWrapper';
@@ -54,7 +50,7 @@ import { ScanTypeEnum, VulnerabilitySeverityType } from '@/types/common';
 import { get403Message } from '@/utils/403';
 import { apiWrapper, retryUntilResponseHasValue } from '@/utils/api';
 import { download } from '@/utils/download';
-import { SeverityEnumList } from '@/utils/scan';
+import { SeverityEnumList } from '@/utils/enum';
 
 const DEFAULT_PAGE_SIZE = 10;
