+++ deepfence_frontend/apps/dashboard/src/features/vulnerabilities/pages/UniqueVulnerabilities.tsx
@@ -35,8 +35,8 @@ import { FilterWrapper } from '@/features/common/FilterWrapper';
 import { queries } from '@/queries';
 import { useTheme } from '@/theme/ThemeContext';
 import { ScanTypeEnum, VulnerabilitySeverityType } from '@/types/common';
+import { SeverityEnumList } from '@/utils/enum';
 import { getOrderFromSearchParams, useSortingState } from '@/utils/table';
-import { SeverityEnumList } from '@/utils/scan';
 
 const DEFAULT_PAGE_SIZE = 10;
