+++ deepfence_frontend/apps/dashboard/src/features/vulnerabilities/pages/VulnerabilityScanResults.tsx
@@ -98,6 +98,7 @@ import {
 import { get403Message, getResponseErrors } from '@/utils/403';
 import { apiWrapper } from '@/utils/api';
 import { formatMilliseconds } from '@/utils/date';
+import { SeverityEnum, SeverityEnumList } from '@/utils/enum';
 import { abbreviateNumber } from '@/utils/number';
 import {
   isScanComplete,
@@ -106,8 +107,6 @@ import {
   isScanInProgress,
   isScanStopped,
   isScanStopping,
-  SeverityEnum,
-  SeverityEnumList,
 } from '@/utils/scan';
 import {
   getOrderFromSearchParams,
