+++ deepfence_frontend/apps/dashboard/src/queries/vulnerability.tsx
@@ -21,7 +21,8 @@ import { DF404Error } from '@/components/error/404';
 import { VulnerabilitiesCountsCardData } from '@/features/vulnerabilities/components/landing/VulnerabilitiesCountsCard';
 import { getResponseErrors } from '@/utils/403';
 import { apiWrapper } from '@/utils/api';
-import { SCAN_STATUS_FILTER_TYPE, SCAN_STATUS_GROUPS, SeverityEnum } from '@/utils/scan';
+import { SeverityEnum } from '@/utils/enum';
+import { SCAN_STATUS_FILTER_TYPE, SCAN_STATUS_GROUPS } from '@/utils/scan';
 
 export const vulnerabilityQueries = createQueryKeys('vulnerability', {
   scanList: (filters: {
