+++ deepfence_frontend/apps/dashboard/src/utils/scan.ts
@@ -1,12 +1,6 @@
 import { generatePath } from 'react-router-dom';
 
-import {
-  ModelMalwareFileSeverityEnum,
-  ModelScanInfoStatusEnum,
-  ModelSecretLevelEnum,
-  ModelVulnerabilityCveSeverityEnum,
-} from '@/api/generated';
-import { ModelBenchmarkType } from '@/api/generated';
+import { ModelScanInfoStatusEnum } from '@/api/generated';
 import { ScanTypeEnum } from '@/types/common';
 
 export const isScanComplete = (status: string): boolean => {
@@ -137,40 +131,3 @@ export const SCAN_STATUS_FILTER = {
   Cancelling: 'Cancelling',
   Deleting: 'Deleting',
 } as const;
-
-export const SeverityEnum = {
-  ...ModelSecretLevelEnum,
-  ...ModelVulnerabilityCveSeverityEnum,
-  ...ModelMalwareFileSeverityEnum,
-} as const;
-
-export const SeverityEnumList = [
-  SeverityEnum.Critical,
-  SeverityEnum.High,
-  SeverityEnum.Medium,
-  SeverityEnum.Low,
-  SeverityEnum.Unknown,
-] as const;
-export function getBenchmarkPrettyName(backendBenchmark: ModelBenchmarkType) {
-  switch (backendBenchmark) {
-    case ModelBenchmarkType.Cis:
-      return 'CIS';
-    case ModelBenchmarkType.Nist:
-      return 'NIST';
-    case ModelBenchmarkType.Pci:
-      return 'PCI';
-    case ModelBenchmarkType.Hipaa:
-      return 'HIPPA';
-    case ModelBenchmarkType.Soc2:
-      return 'SOC2';
-    case ModelBenchmarkType.Gdpr:
-      return 'GDPR';
-    case ModelBenchmarkType.NsaCisa:
-      return 'NSA-CISA';
-
-    default:
-      // eslint-disable-next-line no-case-declarations
-      const _exhaustiveCheck: never = backendBenchmark;
-      throw new Error(`Unhandled case: ${_exhaustiveCheck}`);
-  }
-}
